requirements.txt
Django~=4.2.0
djangorestframework
psycopg2-binary
dj-database-url
gunicorn
whitenoise
djangorestframework-simplejwt
python-dotenv
django-cors-headers
